% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_map_state}
\alias{fars_map_state}
\title{Generate a map showing the number of measurements in a specified state}
\usage{
fars_map_state(state.num, year)
}
\arguments{
\item{state.num}{integer}

\item{year}{integer}
}
\value{
returns a map showing the number of accidents in a state
}
\description{
Generate a map showing the number of measurements in a specified state
}
