% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read_years}
\alias{fars_read_years}
\title{Read all files that correspond to a vector of years}
\usage{
fars_read_years(years)
}
\arguments{
\item{years}{a vector of years}
}
\value{
returns a list of data.frames, one data.frame for each year. Will provide a NULL value and error message for any year where no corresponding file can be found
}
\description{
Read all files that correspond to a vector of years
}
