% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_summarize_years}
\alias{fars_summarize_years}
\title{Collect and summarize data over multiple years}
\usage{
fars_summarize_years(years)
}
\arguments{
\item{years}{vector of years}
}
\value{
summary table showing the count of measurements per month and year
}
\description{
Collect and summarize data over multiple years
}
