% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{make_filename}
\alias{make_filename}
\title{This builds the name of the file based on the year}
\usage{
make_filename(year)
}
\arguments{
\item{year}{integer}
}
\value{
returns a filename concatenating "accident_" with the year and file extensions
}
\description{
This builds the name of the file based on the year
}
\examples{
make_filename(2019)
}
