% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read}
\alias{fars_read}
\title{Read FARS data}
\usage{
fars_read(filename)
}
\arguments{
\item{filename}{this is the filename of the file that will be loaded}
}
\value{
returns a dataframe of the loaded data
}
\description{
This function reads the raw data out and returns a data.frame
}
